Add the ability for mods to put their own items in the top panel.
Added ClickableUIElement for easy use of making your own.
Top panel items are set to be 64x64.
All added items are added to a TopPanelGroup that handles positioning,
rendering, and updating the items.

Note: Currently works but if someone is using modifiers then the added
items will be covered up. Planning on converting modifiers into a
dropdown list at some point. If you don't think it should be accepted
until that is implemented that's fine.