Fix Placement of items on the top panel.
Converted the Daily Mods into a dropdown list.
Make the player name get cut off it its too long.
Change how ascension level is displayed to give more room.
Repositioned the TopPanelGroup to fit with the changing of item positions.

Add methods to BaseMod to add and remove a panel item that is part of
the top panel group.