Event Override Default Functionality
Currently, AddEventParams will have a type of NORMAL if it isn't specified, so if an override event is specified but type is not, it will not override the event at all. This will cause it to default to an OVERRIDE event if type is unspecified but an override event is.